Cleanup of zookeeper initialization
For one, this changes the configuration key to just "zookeeper". The "ip" was
already not great because it also contained the port.

But mainly this cleans up initialization and listeners on zookeeper. It now
properly logs connection changes (instead of stating it's connected directly)
and properly formats error messages from zookeeper.