Keep track of configurations per plugin
This allows us to make configuration values namespaced in zookeeper (per plugin
and per configuration) so we don't accidentally override the same keys for
different configurations of the same/different plugins. So our keys in zookeeper
 will now look like this:

     /minecraft/<plugin name>/<configuration name>/<configuration key>

where 'configuration name' is either "DEFAULT" or a name set by the user and
"configuration key" the yaml configuration key (just with "/" instead of "." as
separators).

Additionally, you can now get the synchronized configurations of other plugins.
If that's good or not is up to the user...