Add instance discovery
This provides a more or less simple implementation of instance discovery
 to this plugin. This is currently opt-in via the config and requires a
little more setup. For one, an instance would need a unique name and a
"host" which would describe the host(and ip) the server is running on
that the other instances could use to access it. An example could look
like this:

    announce_instance: true
    instance: Uber
    host: localhost:25565

Upon server start, it would then get registered in zookeeper (along
with a list of plugins) and other servers would notice a new server
instance. The instances are all cached so the information can be
retrieved rather quickly. This can be done the following way:

    syncService.getInstanceWatcher().getAllInstances()

Which would give you a list of instances that are registered (including
this instance). The only implementation detail that I'm not happy about
is the duplication of the child caching which I sadly can't avoid right
now.