Allow for synchronizing data not connected to a configuration
This provides a new "synchronized storage" which can store arbitrary -
although required to be json serializable - data in zookeeper alongside
normal data. It behaves very much the same way as when you're
synchronizing a configuration, the only difference being that there's no
 configuration backing this. Keys can look however you like, but work
similar to how they work in a configuration.

This also adds the ability to, at any point in time, get the current
value of a synchronized key directly, instead of having to ask the
configuration for it. This is more useful for the new storage than for
the synchronized configuration, but is available for both.