Add first implementation of a distributed (player) queue
This allows plugin owners to create a distributed queue that could be
used to queue up player for a game, lobby or party and would, once the
queue contains enough players, bunch them together. Once the queue is
cleared, it can accept new entries and the cycle will begin anew.

Currently this requires one instance to be considered the leader, where
the grouping will happen and the queue will be managed. Therefor, this
leader is not allowed to drop out. Any other server on the network does
not have this requirement and can happily disconnect without disrupting
the inner workings of the queue.